+++ src/domains/base-eurc.ts
@@ -1,0 +1,10 @@
+import { EChains, ETokenAddress } from "../types"
+
+const baseEURC = {
+  name: 'EURC',
+  version: '2',
+  chainId: EChains.base,
+  verifyingContract: ETokenAddress.eurcBase
+}
+
+export default baseEURC