+++ src/helpers/define-escrow-address.ts
@@ -1,5 +1,9 @@
 import * as configs from '../configs'
-import { TDeploymentType, TTokenType } from '../types'
+import {
+  TDeploymentType,
+  TTokenType,
+  EChains
+} from '../types'
 
 type TDefineEscrowAddress = (
   chainId: number | null,
@@ -17,6 +21,13 @@ const defineEscrowAddress: TDefineEscrowAddress = (
   }
 
   if (deployment === 'CBW') {
+    if (tokenType === 'ERC1155' || tokenType === 'ERC721') {
+      const escrow = configs.cbwEscrowContractNFT
+      if (!escrow) {
+        return null
+      }
+      return escrow
+    }
     const escrow = configs.cbwEscrowContract
     if (!escrow) {
       return null
@@ -25,7 +36,14 @@ const defineEscrowAddress: TDefineEscrowAddress = (
   }
 
   if (tokenType === 'ERC1155' || tokenType === 'ERC721') {
+    if (chainId === EChains.immutableZkevm) {
+      return configs.immutableZkevmContractNFT
+    }
     return configs.mainEscrowContractNFT
+  }
+
+  if (chainId === EChains.immutableZkevm) {
+    return configs.immutableZkevmContract
   }
   
   return configs.mainEscrowContract